fix(ATE-595): report RDP drive capacity without 32-bit overflow
gcsfuse advertises 2^33 allocation units. Storing that count in a 32-bit
int made Windows Explorer see zero free space and refuse copies onto the
redirected drive. Keep the existing RDPDR UINT64 reply and scale huge
counts into a larger block size so ordinary disks are unchanged.

+++ src/protocols/rdp/fs.h
@@ -231,13 +231,18 @@ typedef struct guac_rdp_fs_info {
 
     /**
      * The number of free blocks available.
+     *
+     * Stored as 64-bit to match RDPDR FileFsSizeInformation. Values that
+     * would not fit in a signed 32-bit count are scaled in
+     * guac_rdp_fs_get_info() so older 32-bit consumers still see non-zero
+     * free space.
      */
-    int blocks_available;
+    uint64_t blocks_available;
 
     /**
      * The number of blocks in the filesystem.
      */
-    int blocks_total;
+    uint64_t blocks_total;
 
     /**
      * The number of bytes per block.
